Made RuleError its own class
