+++ modAionImpl/src/org/aion/zero/impl/valid/AionExtraDataRule.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.aion.mcf.blockchain.BlockHeader;
 import org.aion.mcf.blockchain.valid.BlockHeaderRule;
+import org.aion.mcf.blockchain.valid.RuleError;
 
 public class AionExtraDataRule extends BlockHeaderRule {
