+++ modAionImpl/src/org/aion/zero/impl/valid/AionHeaderVersionRule.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.aion.mcf.blockchain.BlockHeader;
 import org.aion.mcf.blockchain.valid.BlockHeaderRule;
+import org.aion.mcf.blockchain.valid.RuleError;
 import org.aion.zero.impl.types.A0BlockHeaderVersion;
 
 public class AionHeaderVersionRule extends BlockHeaderRule {
