+++ modAionImpl/src/org/aion/zero/impl/valid/EquihashSolutionRule.java
@@ -4,9 +4,10 @@ import java.util.List;
 import org.aion.equihash.OptimizedEquiValidator;
 import org.aion.mcf.blockchain.BlockHeader;
 import org.aion.mcf.blockchain.valid.BlockHeaderRule;
+import org.aion.mcf.blockchain.valid.RuleError;
 import org.aion.zero.impl.types.A0BlockHeader;
 
-/** Checks if {@link A0BlockHeader#solution} is a valid Equihash solution. */
+/** Checks if {@link A0BlockHeader#getSolution()} is a valid Equihash solution. */
 public class EquihashSolutionRule extends BlockHeaderRule {
 
     private OptimizedEquiValidator validator;
