--- modMcf/src/org/aion/mcf/blockchain/valid/IValidRule.java
@@ -1,14 +1,0 @@
-package org.aion.mcf.blockchain.valid;
-
-/** Valication rules interface. */
-public interface IValidRule {
-    class RuleError {
-        public final Class<?> errorClass;
-        public final String error;
-
-        public RuleError(Class<?> errorClass, String error) {
-            this.errorClass = errorClass;
-            this.error = error;
-        }
-    }
-}
