+++ modMcf/src/org/aion/mcf/valid/BlockNumberRule.java
@@ -2,6 +2,7 @@ package org.aion.mcf.valid;
 
 import java.util.List;
 import org.aion.mcf.blockchain.BlockHeader;
+import org.aion.mcf.blockchain.valid.RuleError;
 
 public class BlockNumberRule extends DependentBlockHeaderRule {
